parsers/requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
future==0.18.2
idna==2.10
python-dotenv==0.17.0
requests==2.25.1
soupsieve==2.2.1
urllib3==1.26.4
